requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing & Storage
pandas==2.1.4
numpy==1.26.3
influxdb-client==1.39.0
redis==5.0.1

# Machine Learning & Anomaly Detection
scikit-learn==1.4.0
tensorflow==2.15.0
pyod==1.1.3
prophet==1.1.5

# Monitoring & Metrics
prometheus-client==0.19.0

# Messaging & Alerts
pika==1.3.2
requests==2.31.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-json-logger==2.0.7
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Security & Cryptography
cryptography==41.0.7
